Revert "Refactor path utility functions into utils.js and update watcher.js to use the new module"
This reverts commit 79256c51f432c7984ca7ae7723a0874f591a19ba.
